adjusting import path based on moving action block container
